refactor(st-clock): data size optimization
Reduce type of gate_refcounts[], uint8_t is enough.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/323780
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>

+++ drivers/st/clk/clk-stm32-core.h
@@ -73,7 +73,7 @@ struct stm32_clk_priv {
 	const uint32_t nb_div;
 	struct clk_oscillator_data *osci_data;
 	const uint32_t nb_osci_data;
-	uint32_t *gate_refcounts;
+	uint8_t *gate_refcounts;
 	void *pdata;
 };
